🚚 Rename src/assets folder to src/lottie
