+++ my-code/frontend/src/lottie/loading-animation.json
